Do not apply chains in bridge for vlan traffic
This is a quick workaround to problems like NAT rules on bridges that
require access to fields in higher layers than the bridge would
theoretically be able to see. This would work up to now with TCP/UDP
traffic, but started being problematic on ICMP traffic because NATing on
it requires the extra userspace only fields in the enriched
FlowKeyICMPs. But to get these we need to dig in the VLAN encap, and
this is not quite correct. The bridge should just let vlan-traffic
through but not see inside.

Adding a new document with behaviour information related to Bridges
where this information is documented. I agreed with Pino that this is
the best option over putting the info in the REST API doc where it
doesn't really belong.

Fixes MN-590
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: If5768bea006b191301c08ffad5fd300c0c019c11