Add mm-ctl script
mm-script is a simple script that currently only does port
binding and unbinding.  The reason for the script is so that the
orchesteration system like OpenStack can perform port binding
without needing to make API calls in their code.  Instead, it
would run this command that is installed on the hypervisor host
to simplify the process.

Change-Id: I0eeff02946ac60b99adc0155f0f182ca970f56c8
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>