Adds a check if a config file exists on Midolman start-up.
Adds a check if a Miodlman config file exists at the specified
location. If it doesn't, logs that the file is missing and exits with
an error code of 1. Currently Midolman does not check the file on
start-up, and it receives null references to a config object on
dependency injection, resulting in a null pointer exception at
run-time.

Fix MN-101

Change-Id: I0b7d40403da145505bf042c3521fd10ef2e55021