Refactor probes in midolman simulation tests
This patch refactors probe code in midolman simulation tests to prepare
for further refactors and test revamps.

Various duplications in test helper methods are eliminated as well.

It loosen BridgeInvalidationTestCase#testLeanMAC2() to make it more
robust.

The duration of receiveOneMsg in drainProbes() is set to 0 as it seems
this method sometime "eats" msgs in probes even after it returns,
causing some unit tests to fail randomly. Apparently this may solve
problems with various flakky tests in BridgeSimulationTestCase.

Change-Id: I3187680e2ba2abce784c06b927cd209169ac7d71
Signed-off-by: hugobenichi <hugo@midokura.com>