Framework for single-actor tests
This commit introduces SingleActorTest a new base class for test cases
where all of midolman's top-level actors are dummy MessageAccumulator
instances. These are wrapped by TestActorRefs, so the underlaying instance
can be queried for messages it has seen.

Subclasses will typically start a single actor in their beforeTest() method,
and the test cases will evaluate how that actor responds to external messages,
be it by sending replies to the sender or by interacting with other actors
or mocks.

Change-Id: Ie1a65c09e3fcbdea9933497937b25252ddec40d3
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>

sq

Change-Id: Iee16e233d0d663b37569253ce372819420c9d8c1
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>