Add DeduplicationActorTestCase, and fix a bug
See MN-816. The DDA was, up to now, pending the first packet of each flow
and then treating the set of pended packets as a list where it avoided
emitting the head. The reason for keeping this around, it appears, was
using that packet as an index to clean up dpMatchToCookie. This was wrong
because the pended packets aren't a list, they are a set.

We fix this by not pending the first packet and keeping a reverse map
of cookies to flow matches.

Along with the fix, this patch includes complete unit tests for the DDA.

Fixes: MN-498
Refs: MN-816

Change-Id: I63e0d6e5349a9eb716efb9d248c1426f791ec96f
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>