Fixing expect(WildcardFlowAdded) in unit tests
I noticed that all remaining flaky unit tests are almost always
failing because of missed WildcardFlowAdd events from the
FlowController. This is because the FlowController has logic to avoid
adding wildcard flows when the associated tags are not relevant
anymore. Many unit tests do not take this into account, which leads
to possible failures.

I changed IPFragmentationTestCase, L2FilteringTestCase and
BridgeInvalidationTestCase to now listen to AddWildcardFlow events
instead. This patch should reduce the failure rate of unit tests.
I believe similar changes is all that is needed for virtually all
simulation tests to be able to assert:
    1) that the simulation layer requests adding wildcardflows
    2) that the requested wildcardflows have flow actions matching
        something specific (tunnel actions, drop actions, ...)

Change-Id: Ib9052580ee15a856aeffbc1897188019496e5007
Signed-off-by: hugobenichi <hugo@midokura.com>