SBT compiler for all midonet
Because of a conflict between ClientTest and TestTenant, the pom files
had to set the compilation of midonet-api with the default maven
compiler instead of using SBT. The exact problem was that sbt compiled
test classes would be ran by maven-surefire on jenkins in a different
order that when compiled with maven-compiler. This is why this conflict
was not revealed until now.

After checking test coverage with and without, we are removing the class
ClientTest which was causing problems. This allows us in turn to roll
out sbt over all midonet.

Change-Id: Ib09ce25c5003f17682e2e770a553f809fd5865dc
Signed-off-by: hugobenichi <hugo@midokura.com>