Make WeakObjectPool thread safe
The WeakObjectPool was backed by a thread unsafe WeakHashMap that
could have its internal data structures corrupt specially as a
consequence of lots of collisions. As reported in the bug associated
with this commit, various threads might get trapped in a loop when
querying the WeakHashMap.

This solution uses CustomConcurrentHashMap from the Netty libraries as
the backing data structure of the pool. Not using the version from the
Guava libraries because it's private and hidden behind a factory class
that imposes some implementation details that don't work for us, namely
identity based key comparison.

Fixes MN-858
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ie2aaaf930b5500a75b6f2be86e1bb043aa1089d1