Fixing a race condition in IPFragmentationTestCase
It seems that hard-coded assumptions about vm to port number mapping
in the datapath can be wrong because of concurrency in the tests.
This can in turn cause tests in IPFragmentation to fail because of the
assertions in VMsBehindROuterFIxture#expectPacketOut(). Removing these
assertions as they are not really related to the targets of the tests.

This should fix the failures in jenkins with msg
    "Set(4) did not contain element 3"

Change-Id: Id3969259b0d66ad29fed20faa500a2b8f37020cf
Signed-off-by: hugobenichi <hugo@midokura.com>