Race condition in DatapathFlowInvalidationTestCase
DatapathFlowInvalidationTestCase#testTunnelPortAddedRemoved was
failing from time to time in jenkins because of race condition in the
test.

Assertion of new tunnel route after deleting an old tunnel route was
done right after queuing the change in DataClient, which could cause
the test to fail if the test thread was too fast. Swapped the route
assertion after the flowInvalidation assertion, which should fix the
test.

Change-Id: Ibac6072bca1e21c6b46bec63815daaa0142bd3f7
Signed-off-by: hugobenichi <hugo@midokura.com>