VTPM resilient against VTA request timeouts
If for whatever reason the request of the VTPM to the VTA for port
config fails or timeouts, the VTPM is left in an unconsistent state
and PortSet data are never collected again until the binding of some
local vport associated to the corrupted PortSet changes again.

This patch changes the VTPM to retry after a request timeout. The
retry happens maximum 3 times and only after TimeoutException.

The VTPM will only process a single LocalPortActive msg at a time per
vport, and have only one request to the VTA at most at a time related
to this vport. The VTPM will keep further LocalPortActive msgs  for
later processing after the request to the VTA either succeeds or fails.

Fix MN-880
Signed-off-by: hugobenichi <hugo@midokura.com>

Change-Id: Id018af91abc64f34f46974f719785fe3f77dfd94