Remove icmp_seq from flow key and bugfixes
Because we don't need it anymore, and not using it to calculate matches
will avoid having to deal with parallel processing of ICMPs with the
same id. Also, it fixes the issue below which was caused ultimately
because the icmp_seq in pended packets was being overwritten by the
simulated one.

This also fixes a bug in the WildcardMatch hashCode and toString calculations
where a break was missing in some case clauses.

Includes unit test in PingTest to verify that upon sending a bunch of
icmps with the same id, but different seqs, the pended packets get sent
correctly.

Change-Id: Icaf42f7af23c87a5eed2b85acee14e6674134684
Fixes: MN-273
Ref: MN-883, MN-879
Signed-off-by: Galo Navarro <galo@midokura.com>