Modular VTPM for easier unit testing
Breaking @Inject dependent code into subfunctions exposed to the VTPM
via mixin traits. For the time being these traits are mixed statically
directly in the VTPM class, but can be mixed at Actor instantiation in
the MidolmanActorService module, or replaced in the the test by mock
traits in a later patch.

Ref MN-495
Signed-off-by: hugobenichi <hugo@midokura.com>

Change-Id: Ic0f89233b5d5f10175c854462079dbd340f55c81