Prevent cast exception in FlowMatches
Wrong assumptions in a cast were throwing exceptions for ICMP packets
other than echos, unreachable, time_exceeded and parameter_problem. This
fix refactors this so that the only cast required is much safer since
it'll cast from FlowKey.UserspaceOnly to FlowKey<?> which should always
be safe. But not from a FlowKey<?> to FlowKey.UserspaceOnly, which is
not safe.

Add a test case with a ICMP type that does not generate a UserspaceOnly
FlowKey to verify that this works.

Fixes MN-900
Ref MN-879
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I566005e047eaf7f807060e3d4916edf82565d3f6