PacketWorkflow: run cleanup code on failure
A relatively recent change to the PacketWorkflow, separated the
failure exit code path from the general one so that exceptions
could be logged.

However it failed to keep the necessary exit cleanups. The consequence
is that a failed workflow will not release the simulation throttling
token, leaking them.

Fixes MN-901
Refs MN-861

Change-Id: If2a88d528e7c71b0616bf649912ae3d2206347fa
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>