PacketContext: turn cookies into options
Prior to this patch, PacketContext was receiving an Option[Int]
on construction time as the cookie. However, its public API was
peeling off the option and returning null or Integer to callers.

This API created unnecessary boilerplate code in most of the callers
and allowed a null value to creep into the wrong place in MN-887.

This patch moves the public API to be option based, porting the
parent interface to Scala too.

Some unused classes have been deleted in the process.

Fixes: MN-887.

Change-Id: Ica9a3de2f5885ae2241d33de953690569e7baf72
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>