Move throttling token release to DDA
This patch moves the throttling token release to the DDA so it can be
integrated with the cookie expiration mechanism.

This prevents the leak that would happen if a simulation fails to finish
or fail properly for some reason. Also, it hides the token release
inside the DDA, whereas before it was spread out in two different classes.

Related to MN-867

Change-Id: Ib80bf54a7aa3f1ce3d389cf8f4da6e891f97ca0d
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>