Refactored VirtualTopologyActor::expiringAsk
This patch introduces a refactoring of the expiringAsk method. It now
returns either a completed future holding the requested device of the
type specified by the request object, or a failed future representing
some error. Logging of the failures is now encapsulated within the
method.

Furthermore, this patch replaces the Everything case class with a map
from UUID to Device.

General revamping of error handling across callers of expiringAsk.

Change-Id: If6128e4928c2dd45cb0d0f609316118f8783a309