PacketWorkflow isolation from Coordinator
In the context of unit testing the PacketWorkflow, this patch uncouples
the PacketWorflow from the actual simulation layer.

It introduces a factory method in the PacketWorkflow object which takes
as argument a curried function that runs a simulation and returns a
SimulationResult.

The function can be supplied by the Deduplication Actor to create a
Coordinator object to call the simulate method on it when it
instantiates a new PacketWorkflow instance.

With this api, a PacketWorkflow can be tested with predefined
Simulation behavior without an actual Simulation layer.

Ref MN-496
Signed-off-by: hugobenichi <hugo@midokura.com>
Change-Id: I30f754d45369d86956722057992571abbb05522e