DDA: ability to 'pause' simulations
The DDA accepts messages that request that a simulation be 'paused'
until a certain Promise is completed. 'Pausing' a simulation means
temporarily releasing its simulation throttling token, allowing others
to take its place. A ring buffer limits the number of simultaneously
paused simulations, when the ring buffer is found to be full, the oldest
waiting promise is timed out.

Refs MN-867

Change-Id: I5ca1a48a63baee71e4a8673e6d76a866f9e91e57
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>