Allow changing jar location for mm-*ctl scripts
Previously mm-ctl and mm-dpctl set $MIDO_HOME, the location
of the midolman jars, whether it was set in the environment
or not.

With this patch, the mm*-ctl scripts only change the environment
variable if it doesn't exist.

This allows devstack to point the scripts at a source directory,
which lets us run devstack on a machine without disturbing an
existing MidoNet install on that machine.

Change-Id: Ieb5a16da7ed7bd1b99930b67e180d218d3fd7aad