Setting scala-lang dep to runtime in midonet-util
midonet-util pom had the scala-lang dep wrongly declared as test
instead of runtime. Before the relocation of WeakObjectPool from
packets to midonet-util, the problem was "solved" by the fact that
midonet-util was runtime dependent on packets, and would therefore
get the scala-lang at runtime no matter what. Fixing by removing
the test declaration in midonet-util/pom.xml

Thx to Brandon for spotting this.

Change-Id: I73001e5282958bba70024da8885c8a21afb93392
Signed-off-by: hugobenichi <hugo@midokura.com>