Add new ZK and API for Pool, Member, HealthMonitor
This change adds the ZK and API objects required for Pool, Member, and
HealthMonitor. The changes include:

* New ZK managers
* New API resource handlers
* New DTO objects.
* New UT.

These changes DO NOT include

* FULL UT
* Documentation Changes
* VIP API and ZK objects.

The reason this is not complete is because I expect the content of these
objects to change throughout the course of the project.

Conflicts:

	midolman/src/main/java/org/midonet/cluster/DataClient.java

[Merger (Brandon Berg)]:
Re-adding InvalidStateOperationException.java, which had been deleted
from master.

Change-Id: I7188793c703e1d0334f88531fa8896feee9e7d81