Removed the duplicated SIGTERM handler in HostAgent.java
In HostAgent.java we have the duplicated `SIGTERM` handlers. One is
using `sun.misc.Signal` and `sun.misc.SignalHandler`, and another one is
using `Runtime.addShutdownHook`. They both try to handle `SIGTERM`
spawning the different threads and that causes the problems because of
the conflicts in the finalising process.

This patch elimites the duplicated `SIGTERM` handler especially the one
which is using `sun.misc.Signal` and `sun.misc.SignalHandler` because
they are officially deprecated. See the following documentations for
more detail information:

http://www.oracle.com/technetwork/java/faq-sun-packages-142232.html
http://docs.oracle.com/javase/7/docs/api/java/lang/Runtime.html#addShutdownHook%28java.lang.Thread%29

Related with MN-926.

Change-Id: I4c84a6c1a61b20eea6b34cc67db28bce4ea0791b
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>