Replace the duplicated mappings with the redundant classes
Jackson only supports one-to-one mappings between type names and
`JsonSubType`s. So we created duplicated classes with the different
names, and this patch replaces those duplicated mappings with the
mappings with the appropriate redundant classes.

This patch eliminates the gap to upgrade MidoNet from v1.2 to v1.3.

This fixes MN-955.

Change-Id: I9fc058d2dc4405092db24fd48a55169924267b79
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>