Add package script and update client submodule
Simplifying some of the packaging and tagging tasks. Depends on changes
in the python-midonetclient that will remove debian changelog and make
version in deb and rhel builds parametrizable, as well as a
self-sufficient script inside the submodule.

The packaging script contains instructions on usage, on this commit it
is intended to provice package functionality for both debian and rhel.
Additional functionality is also available to perform tagging and
version update in pom files for final, rc and hotfix versions. These are
still under development.

Change-Id: Ic855a0763d7c9a137a2b5ad6f420166630e8651e
Signed-off-by: Galo Navarro <galo@midokura.com>