Removing Port options setter and getter
Preparations for removing PortOptions parameter in the Port class.

Instead of blindly deserializing port Options from NetlinkMessage,
a new supportOptions() method tells the deserializer if an option
object needs to be created. If so, it then deleguates to the method
setOptionsFrom(), by default a noop. Only the VxLan port subclass
overrides this noop method to actually build a VxLanTunnelPortOptions.

Change-Id: Ib68b9694bb38fd26f513bb9c32981418246ff49b
Signed-off-by: hugobenichi <hugo@midokura.com>