Add VIP APIs for L4LB
This patch adds the CRUD APIs for the virtual IP address resources
(VIPs) of L4LB. VIPs are associate with Load Balancers and Pools in the
many-to-one and one-to-one mapping behaviours for each other. VIPs heve
references for those resources but they're individual resources and
decoupled from each other.

A VIP is a psuedo IP address which works as a proxy for the real IP
addresses of the pool members. When the user hits a VIP, it would load
balance the request with some load balancing algorithms such as round
robin. A load balancer can have multiple VIPs. The current API
specification only describes about the back references from the VIP to
the load balancer and this patch follows the specification.

This partially fixes MN-1033.

Change-Id: I73c266a837a6eb152e5fbd0f882a08715f0534ea
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>