Odl: minimal Zookeeper guice profile
A new class ZKCluster is added to org.midonet.midolman.
This class has a static method startCluster() which loads
the necessary midolman components to start a functional
zookeeper connection and have access to an implementation
of DataClient. This method takes as an input parameter a
local path to a midolman configuration file and is threadsafe.
It us guaranteed that the cluster is started at most once.
Once the cluster has been started, a DataClient instance can
be obtained by calling the static method getClusterClient().

To test this feature directly on the command line, call:
$ java -cp path/to/bundle_jar org.midonet.midolman.ZkCluster

bundle_jar is a jar that contains all internal and external
midolman dependencies. It can be created by maven by calling
the profile "fatJar" with the goal "package":
$ mvn package -DskipTests -am -pl midolman -PfatJar
The jar is put in midolman/target

Change-Id: Ic165883bff80127fa545f89632f0638457a8064a
Signed-off-by: hugobenichi <hugo@midokura.com>