Netlink/Ovs: RequestContext for sending NL msgs
For every netlink message send by the OvsDatapathConnectionImpl,
an expensive 72B RequestBuilder object is instantiated, for the only
purpose of storing arguments before calling sendNetLinkMessage().

To avoid the creation of this object, an overloaded
sendNetlinkMessage() is added to AbstractDatapathConnection
to directly pass in necessary arguments. Some arguments are grouped
together in a NetlinkRequestContext object, that can be prepared
statically by command families.

This patch shows an example with the Datapath family. Because at
startup the system does not know the command family id number, the
NetlinkRequestContext objects for the different crud commands are
created as instance fields of the DatapathFamily class. After the
OvsDatapathConnection has quered the datapath family id and
created a DatapathFamily, the Context objects are available for
sending request to the datapth.

Change-Id: Ief9c080763457b058f1b4d719be747b8b04e47d0
Signed-off-by: hugobenichi <hugo@midokura.com>