Makes FlowActions and FlowKeys immutable
FlowActions and FlowKeys are now immutable, modulo deserialization.

This patch solves a bug that prevented object pooling from working
for almost all types of FlowKeys. After being created and tested
against the object pool, most FlowKeys were mutated. This means
that a FlowKey would not match any of those in the object pool and
would be inserted, but after setting additional fields it would
become equal to all the other homologous objects in the pool.

Addresses MN-1020
Signed-off-by: Duarte Nunes <duarte@midokura.com>

Change-Id: I30527a9302e8531f926f0cb37c5f949cb4b03123