UserspaceFlowActionTranslator is now static
and not a trait anymore. The DDa and PacketWorkflow objects directly
call the static function in it to alter their packets with icmp data
and filter FlowAction lists from ICMPError and ICMPEcho flow keys.
Also, this filtering now does not create local anonymous function and
does not induce extra collection coyping from java to scala to java.

Change-Id: I6654e5998378429527fef4c7789cd0570388b2a7
Signed-off-by: hugobenichi <hugo@midokura.com>