Add the delegated handlers for pools and VIPs to load balancer handlers.
This patch adds the GET and POST handlers for the following URIs.

* /load_balancers/:loadBalancerId/pools
* /load_balancers/:loadBalancerId/vips

In addition to that this patch adds the cascaded deletion of the VIPs
and the pool members associated with the pools, which would be also
deleted by cascading when the associated load balancers are deleted.

This partially fixes MN-1148.

Change-Id: I3dcf9a7bb36974452cf250b43202d0a8ae7e7964
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>