Generalizes the idea of pre-fetching a device
This patch refactors DeviceManager into a general infrastructure
that allows pre-fetching associated devices before sending the
managed device to the VTA.

Additionally, this patch corrects a bug in DeviceManager, where it
would keep receiving updates from a chain it was no longer interested
in.

It also refactors some misleading code in the RouterManager and
removes some superfluous flow tags.

Change-Id: Ia15667922c0546d36e41c78d5bbfd051f358bd3d
Signed-off-by: Duarte Nunes <duarte@midokura.com>