Throttling: tie throttler tokens to the Packet class
This removes chances of leaking tokens, and makes it ok to release a token
twice, as it will have no effect. This makes the logic easier to follow,
the invariant now is that every packet must have at least one releaseToken()
call made on it, more calls are ok.

Extra calls have been added before making requests to the kernel, so the
number of active simulations will be closer to reality, making throttling
easier to tune.

Change-Id: I918b25231b4cd0c533fd61f6ecdfbc3efe410c25
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>