Changes the port scan tool from nmap to mz.
Changes the performance testing script to use the mz command instead
of nmap. A caveat is that mz does not have an option for packet rate,
therefore we need to control packet rate by specifying a proper delay
in between packets. Currently a delay (in usec) is calculated simply
as 1,000,000 / RATE as a very rough approximation. In practice, this
does't result in sending RATE packets per second because of the time
it takes mz to create / send a packet.

Fixed packet count and src/dest port specifications.

Fix MN-1158

Change-Id: I632a3a22be889546a5c8d223b29dcfa2f0fa51ea