Release script: do not manage submodules
This is in preparation for further changes in packaging, in upcoming changes we
will remove the need to update version numbers in version-controlled files, and
there will be a self-sufficient package script in the submodule to perform all
required tasks.

The last commit in the updated revision of the python-midonetclient
brings the aforementioned changes.

Change-Id: I76aa854bc9693836f5eec87e48fb927939de028d
Signed-off-by: Galo Navarro <galo@midokura.com>