Merge UnixDomainChannel and UnixDomainChannelImpl
Similarly to the NetlinkChannel & NetlinkChannelImpl merge,
UnixDomainChannel keeps the implementation in a single place while
UnixDomainChannelImpl is here only for the sun nio compatibility.

In addition, a more robust way of instantianting child
UnixDomainChannels for accept() is added, using the
NetlinkSelectorProvider. This avoids exposing to the
compiler directo calls to constructors of classes
implementing SelChImpl.

Change-Id: I4e2d4cfb930e8712940e282a14862063a2137f55
Signed-off-by: hugobenichi <hugo@midokura.com>