Pauseless Sims: add local cache to VTPM
This is a preparation for the removal of Futures from the simulation
path. The VTPM now contains a local cache for all devices that may be
requested, which it will serve immediately upon request in a completed
future. In a miss, an ask will be sent to the VTPM actor. Later patches
will be able to use this to stop a simulation whenever an incompleted
future is found.

Change-Id: I874a6f6d326829843a4004291d3e3d154a40125d
Signed-off-by: Galo Navarro <galo@midokura.com>