Streamlining dpport request in OvsDatapathCon
Similarly to the Datapath class, the DpPort now knows how to write
Netlink message for the various commands of the port family. In
addition DpPort is turned into a "quasi"-immutable class, where
quasi means there is no public instance setters left on the class.
For compatibility with legacy tests and the current deserialization
code, it is however necessary to leave internal fields mutable for
the time being.

Change-Id: Iaca541af416dc502cf9823fd9f23a124823932e6
Signed-off-by: hugobenichi <hugo@midokura.com>