Pauseless simulations.
Using the new Urgent to replace all Futures in the simulation path, we
allow the DDA to park a simulation until the required future completes
and then retry the simulation by restarting the workflow.

The future-waiting mechanism has been encapsulated entirely in the
WaitingRoom introduced in a previous commit, so the DDA only needs to
deal with a new type of PipelinePath returned by the PacketWorkflow.

Includes:
- A fix to the Urgent class: NotYet's Future type is not the same as
  Ready's type.
- Removal of the async get from the ArpCache, it was actually doing a
  sync call but wrapped in a callback.
- Removal of SuspendedPacketQueue.

Change-Id: Ie56630ba134195c56d82b80e148069e8259b4b10
Signed-off-by: Galo Navarro <galo@midokura.com>