Add an embedded Jetty instance in Midolman
This commit introduces a servlet container embedded in Midolman,
which is part of the effort to implement host management dash-
board service for operators to diagnose various problems in
Midolman.

This patch does not contain the dashboard itself. Instead, it
just adds a container and enables Midolman to automatically deploy
the dashboard implemented in WAR, without stopping Midolman.
It also adds a couple of relatively static service (e.g. browsing
logs, config files and JMX metrics) available via HTTP, port 61080
by default. The port was decided to avoid Linux's ephemeral port
range, 32768-61000 by default.

The dashboard is going to be maintained in a separate module,
independent of Midolman. The purpose of this patch is to allow
such isolation.

Change-Id: Ic10e6a4f74b9b7f41937c2f2f4bad7609d8b2b3f
Signed-off-by: Takayuki Usui <takayuki@midokura.com>