Easy access to the Midolman instance and the Guice injector
This allows the host management dashboard to easily inspect
the Midolman instance and other central modules instanciated
by Guice.

The dashboard is implemented in webapp and loaded in running
Midolman without stopping it. The even small changes in the
commit make much easier for the dashboard to diagnose the
internal states of running Midolman.

Change-Id: I3d5c6bbe17e7ae7872fcb43fca19bdb285228d0b