VirtualToPhysicalMapper ignores port set request for non-cached device.
VirtualToPhysicalMapper currently tries to send back null to the
message sender when receiving a request for a port set for which it
does not have a cached device with a corresponding tunnel key.
However AKKA DOES NOT allow sending null message, thus the above
attempt results in crashing the VTPM.

This patch just lets the VTPM ignore the request and the request
time out.

Fix MN-1377

Change-Id: I96a36cace1e1d83979950d0d760adaa304650c3c