Install kernel flows for icmp not traversing NAT
Icmp processing is done in userspace by default because the datapath
cannot match some icmp fields that midolman uses for flow programming.
However, this does not concern all kind of icmp processing for all
kind of virtual topologies. Especially, icmp traffic which has not
traversed any NAT rules can be handled in the datapath.

This patch optimizes icmp processing when a simulation return a packet
context with a modified wildcard match whose icmp_id and icmp_data
fields has not been read. To do so, when the Coordinator pushes back
an AddVirtualWilcardFlow msg to the PacketWorkflow, it now adds the
set of matching fields read by the simulation. This gives the
opportunity to the PacketWorkflow of installing datapath flows more
precisely based on which fields has been used.

Fixes MN-1071
Signed-off-by: hugobenichi <hugo@midokura.com>

Change-Id: I2c213399228ca45e0204412aea7cad6eb8c0e990