Fix flakky FlowsExpirationTestCase
The test code was trying to assert that the FlowController was
correctly removing wildcard flow after a given amount of time.
The logic to test is something like
  deletionTime - creationTime > time_constant

To check this the test code was not using the real deletion and
creation times, but instead taking timestamps shortly after the
creation and deletion events. However, because of asynchronous flow
of code, the first timestamp can get delayed to the point where
timestampAfterDeletion - timestampAfterCreation can be smaller than
the time constant.

To fix this, the creation timestamp is taken *before* the creation
event, which ensures that:
    timestamAfterDeletion - timestampBeforeCreation >
        deletionTime - creationTime

Change-Id: Ie287e9cf282802e1495cb584a631d5df5eb92600
Signed-off-by: hugobenichi <hugo@midokura.com>