Fixes memory leak when expiring packets
This patch fixes a memory leak that left pending packets around when
we expired the simulated packet. It also refines HTB integration to
account for packets that are re-simulated.

An issue this patch does not address is adding a bucket for the
waiting room, to make sure we cap the number of restarted simulations.

Ref MN-1455
Signed-off-by: Duarte Nunes <duarte@midokura.com>
Change-Id: I9b2ece9915770e946bbc802d6493dab611da04a0