Removes the ApplyFlow message
This patch introduces a cache in the DDA of FlowMatch to FlowActions,
which enables these actors to get ahead of the FlowController in terms
of processed packets.

After a simulation, an entry gets written into that cache, which
serves to execute pending packets with the same FlowMatch and also
new matching packets that come in. When the FlowController finishes
adding a new WildcardFlow, it writes the corresponding FlowMatch in a
ring buffer shared with a DDA.

At the end of every simulation, a DDA will check to see if its ring
buffer is full, spinning on it until there is capacity for another
simulation (meaning that the FlowController is catching up).

When the DDA finishes processing a batch of packets, it tries to clear
the FlowMatches that are in the cache and have already been added to
the buffer.
Signed-off-by: Duarte Nunes <duarte@midokura.com>

Change-Id: Id445b437f33364c9537dc9eb379626997656c548