Removing Option boxing of tags in flow translation
The tags are always required for VirtualWildcardFlow translation, but
always unnecessary for translating action before sending packets. To
communicate this, half of the flow translation logic was using Option
while the other part was doing null check.

This patch simplifies the tags handling inside the FlowTranslator by
removing the option wrapping and assuming that the tag set reference is
always non null. When translating actions for sending a packet, a
throw-away set of tags is created for impedance matching.

Change-Id: Ic91738b2941132fae92001e1baddf0651d6247e4
Signed-off-by: hugobenichi <hugo@midokura.com>