Completed workflows leave the waiting room
Fix a regression introduced in 2c88c20d5579b1b81b90385c9fef7f9eb69f1331
Before, the giveUpWorkflows was skipping non-idle workflow returned
from the waiting room. This patch reintroduce a check to only give up
workflows that are not active, and log a warning msg otherwise. In
addition, when a workflow complete, it is removed from the waiting
room.

Fix MN-1517
Signed-off-by: hugobenichi <hugo@midokura.com>

Change-Id: I92c1ee53044ce524cd22a1b6fb2ba33c825560f1