RoutingHandler knows router dp port uplink pid
Before the work on resource protection demultiplexed the unique netlink
channel into request channels and uplink channels for every datapath
ports, the pid of the unique channel was known by the DatapathController.
This information was then used in the FlowTranslator to fix every
FlowActionUserspaces with that unique pid.

That mechanism broke after dp ports got their own uplink channels, and
as a consequence it also affected the handling of bgp. This patch
restores it by making the UpcallDatapathConnectionManager returns the
pid of upcall channels created for new datapath ports. After the
RoutingHandler sends a request to the DatapathController for creating
a dp port that will be linked to the routing daemon, the pid of the
upcall channel for that port is reported back to the RountingHandler,
which can then create WildcardFlow with correct FlowActionUserspaces.

Fix MN-1474
Signed-off-by: hugobenichi <hugo@midokura.com>

Change-Id: I4506dc38eda6213e6280eb2f80bd173af13b66cb