Expire pending packets (more aggressively)
Assume packets P1, P2, ... are coming with a single FlowMatch F.
P1 comes in, and a simulation is postponed for some reasons while
processing it with a cookie C1. Now dpMatchToCookie has a mapping
F to C1 and P1 is in waitingRoom.

Then P2 comes in, but since P2 has the same FlowMatch as P1,
P2 is added to cookieToPendedPackets being associated with C1.
Then P3 comes in, ...

Without this patch, P1 is never expired and all packets with
the FlowMatch F are just added to cookieToPendedPackets,
_unless_ another simulation is postponed with another FlowMatch F'
That triggers postponeOn and removes packets from waitingRoom.

In L2GW functional tests, a simulation is possibly postponed when
truck interface status is down. But, packets may not be handled
even after the interface is up again, because of the issue described
above.

See also
804d886a6723fee013396503128e04f3412197ff
2ce96550d4b5a3866bf0a132d05caa667396cbc8

Related to MN-1483

Change-Id: I332dbac5091c3467f9776136a11f46d6737c4d01
Signed-off-by: Takayuki Usui <takayuki@midokura.com>