Prevent L4LB resources from being updated if mappingStatus is PENDING_*.
This patch introduces an exception thrown when the associated pool's
mapping status is PENDING_CREATE, PENDING_UPDATE or PENDING_DELETE and
it is translated into ServiceUnavailableHttpException with 503 and
`RetryAfter` header, which is set as three seconds by default. This
prevent L4Lb resources from being updated until `mappingStatus` in the
pool is changed by the health monitor.

Change-Id: I8c1e73a4759623f86b5c4e252802b917514ff849
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>