Loosen the test conditions on TestNatLeaseManager
The TestNatLeaseManager was testing that the number of collisions for
a given test was below a hardcoded number. This patch changes the test
to instead compare against a percentage, and set the percentage at
a slightly higher level that the hardcoded number implicit percentage.
This should help to avoid spurious failures in jenkins.

Change-Id: I8012f6909b76a68ac68f743dcc78440aa83eb554
Signed-off-by: hugobenichi <hugo@midokura.com>