Force Pool-HealthMonitor mapping to be disassociated before association
The current API implementation allows users to put a health monitor ID
even if another health monitor is associated with a pool. This behaviour
is inconsistent with LBaaS and causes race conditions where the state
mutations on the association (`PENDING_CREATE` -> `ACTIVE`) and the
disassociation (`PENDING_DELETE` -> `INACTIVE`) would compete with each
other.

So this patch modifes the API to return an error response, 400 Bad
Request, when the pool and the health monitor is already mapped and
users are trying to put another health monitor ID in the pool's
`healthMonitorId`. The users need to put null to disassociate the
mapping first, then another request where the ID of another health
monitor is populated. If `null` is set to `healthMonitorId` and the
pool's `mappingStaus` is `PENDING_*`, it throws
`MappingStatusException` and users get 503 Service Unavailable.

This patch also enables disabled tests to validate this behaviour.

This fixes MN-1566, MN-1568

Change-Id: I335bfa161676adc40ca5dfdcb52e97ca007e4939
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>